ml-inference/requirements.txt
pydantic>=1.10
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
python-multipart>=0.0.6